requirements.txt
psycopg2-binary
SQLAlchemy
pyyaml